moving user interface into userinterface folder
